LayoutTests/canvas/philip/tests/2d.text.draw.fill.unaffected-expected.txt

< [index] >
2d.text.draw.fill.unaffected
fillText does not start a new path or subpath
References: 2d.text.draw.fill
Actual output:
Expected output:

Passed
